+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## (Unreleased)
 
+BREAKING CHANGES:
+
+* data-source/mongodbatlas_federated_settings_identity_provider: The only allowed format for `identity_provider_id` is a 24-hexadecimal digit string ([#2185](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/2185))
+* resource/mongodbatlas_federated_settings_identity_provider: Import can only use a 24-hexadecimal digit string that identifies the IdP, `idp_id`, instead of `okta_idp_id` ([#2185](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/2185))
+
 NOTES:
 
 * provider: New changelog format has been incorporated following [Terraform Changelog Specification](https://developer.hashicorp.com/terraform/plugin/best-practices/versioning#changelog-specification) ([#2124](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/2124))
@@ -19,6 +24,8 @@ ENHANCEMENTS:
 
 BUG FIXES:
 
+* data-source/mongodbatlas_advanced_cluster: Converts `replication_specs` from TypeSet to TypeList. This fixes an issue where some items were not returned in the results. ([#2145](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/2145))
+* data-source/mongodbatlas_advanced_clusters: Converts `replication_specs` from TypeSet to TypeList. This fixes an issue where some items were not returned in the results. ([#2145](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/2145))
 * data-source/mongodbatlas_network_peering: Ensures `accepter_region_name` is set when it is has the same value as the container resource ([#2105](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/2105))
 * resource/mongodbatlas_cluster: Fixes nil pointer dereference if `advanced_configuration` update fails ([#2139](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/2139))
 * resource/mongodbatlas_maintenance_window: Fixes `day_of_week` param as **required** when calling the API ([#2163](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/2163))